+++ packages/codegen-ui-react/lib/forms/form-renderer-helper/typescript-type-map.ts
